% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare-reference.R
\name{compare_reference}
\alias{compare_reference}
\title{Compare Reference}
\usage{
compare_reference(reference = 10, n = 5, mean = 11, sd = 0.84)
}
\arguments{
\item{reference}{a reference population mean}

\item{n}{number of sampels}

\item{mean}{sample mean}

\item{sd}{sample standard deviation}
}
\value{
Returns a list of values.
}
\description{
Generates \code{n} values from a normal distribution with mean (\code{mean})
and standard deviation (\code{sd}).
}
\examples{
compare_reference()
}
\references{
Krzywinski, Martin, and Naomi Altman. "Points of significance:
Comparing samples--part I." Nature Methods 11.3 (2014): 215-216.
}

