% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population-dist.R
\name{distr_to_sample_from}
\alias{distr_to_sample_from}
\title{Effects of Random Distribution Sampling}
\usage{
distr_to_sample_from(samples = 100)
}
\arguments{
\item{samples}{integer; number of samples to take from population
distribution}
}
\value{
N/A
}
\description{
Creates non-normal population distribution to draw samples from and shows
resulting sample means.
}
\examples{
distr_to_sample_from()
}

