% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{inference}
\alias{inference}
\title{Inference Errors}
\usage{
inference(alpha = 0.05, beta = 0.2)
}
\arguments{
\item{alpha}{Type I error or the false positive error}

\item{beta}{Type II error or the false negative error}
}
\value{
list containing the specificity and sensitivity
}
\description{
Returns the specificity and sensitivity values for a given alpha and beta
value.
}
\details{
Specificity is also known as the true negative rate, which is calculated as
the number of true negatives divided by the sum of true negatives and false
positives.

Sensitivity is also known as the true positive rate, which is calculated as
the number of true positives divided by the sum of true positives and false
negatives.
}
\examples{
inference(alpha = 0.05, beta = 0.2)
inference(alpha = 0.001, beta = 0.2)
inference(alpha = 0.05, beta = 0.5)
}
