% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population-dist.R
\name{location_plot}
\alias{location_plot}
\title{Generate Location Plot}
\usage{
location_plot(mu = c(-5, 0, 5), std = 1, n = 100, xmin = -10,
  xmax = 10)
}
\arguments{
\item{mu}{mean}

\item{std}{standard deviation}

\item{n}{sample size}

\item{xmin}{x-axis minimum limit}

\item{xmax}{x-axis maximum limit}
}
\value{
Creates three density plots with different center locations
}
\description{
Creates three density plots with varying means. The means are highlighted
with a vertical line.
}
\examples{
location_plot()
}

