% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population-dist.R
\name{population_dist}
\alias{population_dist}
\title{Generate Population Distribution}
\usage{
population_dist(mu = 0, std = 1, n = 100)
}
\arguments{
\item{mu}{population mean}

\item{std}{population standard deviation}

\item{n}{sample size}
}
\value{
Creates density plot with annotations plus prints out mean and
  standard deviation
}
\description{
Creates a density plot with mean and standard deviation annotations.
}
\details{
This function aims to show that the mean, standard deviation, and sample size
of the distribution can influence the distribution's shape and size.
}
\examples{
population_dist()                # Mean = 0, Stdev = 1, N = 100
population_dist(mu = 2)          # Mean = 2, Stdev = 1, N = 100
population_dist(mu = 1, std = 3) # Mean = 1, Stdev = 3, N = 100
population_dist(std = 3, n = 10) # Mean = 0, Stdev = 3, N = 10
}
\references{
Krzywinski, Martin, and Naomi Altman.
  \href{http://dx.doi.org/10.1038/nmeth.2613}{"Points of significance:
  Importance of being uncertain."} Nature Methods 10.9 (2013): 809-810.
}

