% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppv.R
\name{ppv}
\alias{ppv}
\title{Positive Predictive Value}
\usage{
ppv(effectSize, power, alpha)
}
\arguments{
\item{effectSize}{the quantitative measure of the strength of a phenomenon
in an experiment, between 0 and 1}

\item{power}{probability that the test correctly rejects the null hypothesis
when the alternative hypothesis is true}

\item{alpha}{false positive value or Type I error}
}
\value{
positive predictive value
}
\description{
Calculates positive predictive value given the effect size, power, and alpha
level.
}
\details{
The positive predictive value is the fraction of true positives out of
positive labeled classes (i.e. true positive and false positive).
}
\examples{
ppv(effectSize = 0.5, power = 0.2, alpha = 0.05)
ppv(effectSize = 0.5, power = 0.5, alpha = 0.05)
ppv(effectSize = 0.5, power = 0.8, alpha = 0.05)
}
\references{
Krzywinski, Martin, and Naomi Altman. "Points of significance:
  Power and sample size." Nature Methods 10.12 (2013): 1139-1140.
}

