% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vary-box-size.R
\name{vary_box_size}
\alias{vary_box_size}
\title{Vary Box Plot Sizes}
\usage{
vary_box_size(n = 5, plotNum = 5)
}
\arguments{
\item{n}{sample size}

\item{plotNum}{number of box plots}
}
\value{
Plot box plots of five samples from a normal distribution
}
\description{
Plots \code{plotNum} number of box plots with the same sample size \code{n}
sampled from a normal distribution.
}
\details{
Box plots are very powerful visualization tools. However, you need sufficient
number of samples for them to be confident sources of information.

Try changing the sample size of the plots drawn to understand the variability
in visualizing small datasets.
}
\examples{
vary_box_size()
}
\references{
Krzywinski, Martin, and Naomi Altman. "Points of significance:
  visualizing samples with box plots." Nature Methods 11.2 (2014): 119-120.
}
